+++ protocol-code-generator/src/main/java/dev/cirras/generate/type/StructType.java
@@ -5,11 +5,13 @@ import java.util.Optional;
 public final class StructType implements CustomType {
   private final String name;
   private final Integer size;
+  private final boolean bounded;
   private final String packageName;
 
-  public StructType(String name, Integer size, String packageName) {
+  public StructType(String name, Integer size, boolean bounded, String packageName) {
     this.name = name;
     this.size = size;
+    this.bounded = bounded;
     this.packageName = packageName;
   }
 
@@ -21,6 +23,11 @@ public final class StructType implements CustomType {
   @Override
   public Optional<Integer> getFixedSize() {
     return Optional.ofNullable(size);
+  }
+
+  @Override
+  public boolean isBounded() {
+    return bounded;
   }
 
   @Override